fix int

+++ src/main/java/com/isp392/repository/ToppingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ToppingRepository extends JpaRepository<Topping,Long> {
+public interface ToppingRepository extends JpaRepository<Topping,Integer> {
     boolean existsByname(String name);
 }
